[ANDROSDK-2228] move LegacyDataValueApi back to core
